+++ Zephyr/Modules/Miscellaneous.lua
@@ -1,15 +1,13 @@
 local ZPR = LibStub("AceAddon-3.0"):GetAddon("Zephyr")
 
-local partyFrameTitle = ZPR:NewModule("partyFrameText", "AceEvent-3.0")
+
 local mouseoverRaidManager = ZPR:NewModule("mouseoverRaidManager", "AceEvent-3.0")
 local buffFrameCollapseExpand = ZPR:NewModule("buffFrameCollapseExpand", "AceEvent-3.0")
 local talkingHead = ZPR:NewModule("talkingHead", "AceEvent-3.0")
-
-function partyFrameTitle:OnEnable()
-	if CompactPartyFrame:IsShown() then
-		CompactPartyFrameTitle:SetAlpha(0)
-	end
-end
+local nameplateHPPercent = ZPR:NewModule("nameplateHPPercent", "AceEvent-3.0")
+local nameplateArenaNumbers = ZPR:NewModule("nameplateArenaNumbers", "AceEvent-3.0")
+local partyFrameTitle = ZPR:NewModule("partyFrameTitle", "AceEvent-3.0")
+local customHPFormat = ZPR:NewModule("customHPFormat", "AceEvent-3.0")
 
 function mouseoverRaidManager:OnEnable()
 	local raidManager = CompactRaidFrameManager
@@ -51,5 +49,99 @@ end
 function talkingHead:OnEnable()
 	hooksecurefunc(TalkingHeadFrame, "PlayCurrent", function(self)
 		self:Hide()
+	end)
+end
+
+function nameplateHPPercent:OnEnable()
+	if IsAddOnLoaded('Plater') or IsAddOnLoaded('TidyPlates_ThreatPlates') or IsAddOnLoaded('TidyPlates') or
+		IsAddOnLoaded('Kui_Nameplates') then return end
+
+	if GetCVar("nameplateShowOnlyNames") == "1" then
+		return
+	end
+
+	local function nameplateHealthText(unit, healthBar)
+		if not healthBar.text then
+			healthBar.text = healthBar:CreateFontString(nil, "ARTWORK", nil)
+			healthBar.text:SetPoint("CENTER")
+			healthBar.text:SetFont(STANDARD_TEXT_FONT, 8, 'OUTLINE')
+		else
+			local _, maxHealth = healthBar:GetMinMaxValues()
+			local currentHealth = healthBar:GetValue()
+			healthBar.text:SetText(string.format(math.floor((currentHealth / maxHealth) * 100)) .. "%")
+		end
+	end
+
+	local function nameplateHealthTextFrame(self)
+		local inInstance, instanceType = IsInInstance()
+		if inInstance and not (instanceType == 'arena' or instanceType == 'pvp') then
+			if self:IsForbidden() then return end
+		end
+
+		if self.unit and self.unit:find('nameplate%d') then
+			if self.healthBar and self.unit then
+				if UnitName("player") ~= UnitName(self.unit) then
+					local unit = self.unit
+					local healthBar = self.healthBar
+					nameplateHealthText(unit, healthBar)
+				end
+			end
+		end
+	end
+
+	hooksecurefunc("CompactUnitFrame_UpdateHealthColor", nameplateHealthTextFrame)
+	hooksecurefunc("CompactUnitFrame_UpdateHealth", nameplateHealthTextFrame)
+	hooksecurefunc("CompactUnitFrame_UpdateStatusText", nameplateHealthTextFrame)
+end
+
+function nameplateArenaNumbers:OnEnable()
+
+	if IsAddOnLoaded('Plater') or IsAddOnLoaded('TidyPlates_ThreatPlates') or IsAddOnLoaded('TidyPlates') or
+		IsAddOnLoaded('Kui_Nameplates') then return end
+
+	if GetCVar("nameplateShowOnlyNames") == "1" then
+		return
+	end
+
+	local frame = CreateFrame("Frame")
+	frame:RegisterEvent("ZONE_CHANGED_NEW_AREA")
+	frame:HookScript("OnEvent", function()
+
+		local U = UnitIsUnit
+		hooksecurefunc(
+			"CompactUnitFrame_UpdateName",
+			function(F)
+				if IsActiveBattlefieldArena() and F.unit:find("nameplate") then
+					for i = 1, 5 do
+						if U(F.unit, "arena" .. i) then
+							F.name:SetText(i)
+							F.name:SetTextColor(1, 1, 0)
+							break
+						end
+					end
+				end
+			end)
+	end)
+end
+
+function partyFrameTitle:OnEnable()
+	CompactPartyFrameTitle:SetAlpha(0)
+end
+
+function customHPFormat:OnEnable()
+	hooksecurefunc("TextStatusBar_UpdateTextStringWithValues", function(f, t, v, _, x)
+		-- f == PlayerFrame.PlayerFrameContent.PlayerFrameContentMain.HealthBarArea.HealthBar or
+		-- f == PlayerFrame.PlayerFrameContent.PlayerFrameContentMain.ManaBarArea.ManaBar or
+		-- f == FocusFrame.TargetFrameContent.TargetFrameContentMain.HealthBar then
+		if f == TargetFrame.TargetFrameContent.TargetFrameContentMain.HealthBar then
+			local F, V, z = format, v, 1000
+			Z = z * z
+			v = (v > Z and F("%.2fM", v / Z)) or (v > (z * 10) and F("%.0fk", v / z)) or v
+			if V or x then
+				t:SetText(F("%s", v))
+				if UnitIsDeadOrGhost("target") then t:SetText(F(""))
+				end
+			end
+		end
 	end)
 end